refactor(generate-g3): extract ingest and aggregation helpers
